requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
APScheduler~=3.11.0
gspread~=6.2.1
oauth2client~=4.1.3
pytz~=2025.2